requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
google-generativeai